% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_div_metrics.R
\name{calc_div_metrics}
\alias{calc_div_metrics}
\title{Calculate diversity and divergence}
\usage{
calc_div_metrics(aln, founder, gen)
}
\arguments{
\item{aln}{DNA alignment in \verb{[ape::DNAbin]} format}

\item{founder}{Name of the founder sequence in the alignment}

\item{gen}{Vector that indicates the generation of each sequence in the alignment,
assumed to be in the same order as the alignment}
}
\value{
tibble including mean divergence for each generation
}
\description{
See \verb{[vignette('analyze_output')]} for more details.
}
\examples{
# This example is somewhat contrived, but it shows how it works.
calc_div_metrics(hxb2_cons_founder, 'B.FR.83.HXB2_LAI_IIIB_BRU.K03455', c(1,2,2))
}
