% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tr_summary_stats.R
\name{calc_int_over_ext}
\alias{calc_int_over_ext}
\title{Calculate mean internal/external branch length ratio}
\usage{
calc_int_over_ext(tr)
}
\arguments{
\item{tr}{Phylogeny}
}
\value{
Mean internal/external branch length ratio
}
\description{
Calculate mean internal/external branch length ratio
}
\examples{
tr <- ape::rtree(100)
calc_int_over_ext(tr)
}
