% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_nt_sub_probs.R
\name{calc_nt_sub_probs}
\alias{calc_nt_sub_probs}
\title{Calculate nucleotide substitution probabilities}
\usage{
calc_nt_sub_probs(aln, tr = NULL, model = "GTR+R(4)+I", rearrangement = "none")
}
\arguments{
\item{aln}{An alignment of class ape::DNAbin}

\item{tr}{Starting tree}

\item{model}{A string providing model (e.g. "GTR+G(4)+I")}

\item{rearrangement}{Type of tree rearrangements to perform, one of
"none", "NNI", "stochastic" or "ratchet" (default: "none")}
}
\value{
Matrix of nucleotide substitution probabilities.
Columns are from and rows are to.
}
\description{
Fr this, you should probably use a set of within-host sequences for the genome region of interest.
}
\examples{
# NOTE: This is just an example.
# (**TODO: ADD AN ACTUAL WITHIN-HOST HIV ENV ALIGNMENT AS AN EXAMPLE?**)
calc_nt_sub_probs(hiv_env_flt_2021)
}
