% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_nt_subst_probs.R
\name{calc_nt_subst_probs}
\alias{calc_nt_subst_probs}
\title{Calculate nucleotide substitution probabilities}
\usage{
calc_nt_subst_probs(
  aln,
  tr = NULL,
  model = "GTR+R(4)+I",
  rearrangement = "none"
)
}
\arguments{
\item{aln}{An alignment of class ape::DNAbin}

\item{tr}{Starting tree}

\item{model}{A string providing model (e.g. "GTR+G(4)+I")}

\item{rearrangement}{Type of tree rearrangements to perform, one of "none", "NNI", "stochastic" or "ratchet" (default: "none")}
}
\value{
Matrix of nucleotide substitution probabilities.
Columns are from and rows are to.
}
\description{
Calculate nucleotide substitution probabilities
}
\examples{
# NOTE: This is just an example. You probably don't want to use 10 random samples.
hiv_env_flt_2021 <- ape::as.matrix.DNAbin(hiv_env_flt_2021)
calc_nt_subst_probs(hiv_env_flt_2021[1:10,])
}
