% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tr_summary_stats.R
\name{calc_parsimony}
\alias{calc_parsimony}
\title{Calculate tree parsimony score for sampling times}
\usage{
calc_parsimony(tr, timepoints)
}
\arguments{
\item{tr}{Rooted phylogeny}

\item{timepoints}{Vector of time points named by tree tip label. All tip
lables must be included in this vector. If you want to exclude certain tips,
you must drop them from the tree prior to using this function.}
}
\value{
Parsimony score of the tree based on the timepoints
}
\description{
Calculate tree parsimony score for sampling times
}
\examples{
tr <- ape::rtree(100)
times <- sample(3, 100, replace = TRUE)
names(times) <- tr$tip.label
calc_parsimony(tr, times)
}
