% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tr_summary_stats.R
\name{calc_prop_survived}
\alias{calc_prop_survived}
\title{Calculate percent of lineages that survived from one generation to the next}
\usage{
calc_prop_survived(tr, timepoints)
}
\arguments{
\item{tr}{Phylogeny}

\item{timepoints}{Vector of time points named by tree tip label}
}
\value{
Tibble including:
\itemize{
\item \code{n_seqs_timepoint}: Number of sequences included in that time point
\item \code{n_clusters}: Number of monophyletic clusters for that time point
\item \code{prop_survived}: Proportion of lineages that survived from the previous generation
}
}
\description{
Calculate percent of lineages that survived from one generation to the next
}
\examples{
tr <- ape::rtree(100)
times <- sample(3, 100, replace = TRUE)
names(times) <- tr$tip.label
calc_prop_survived(tr, times)
}
