% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{calc_q_from_rates}
\alias{calc_q_from_rates}
\title{Calculate Q matrix from nucleotide substitution rates}
\usage{
calc_q_from_rates(rates, mut_rate, generation_time = NULL)
}
\arguments{
\item{rates}{4x4 matrix of individual nucleotide substitution rates for each
potential substitution (per-site per-generation or per-site per-day) with
the following row and column names: A,C,G,T.}

\item{mut_rate}{Overall per-site per-generation mutation rate.}

\item{generation_time}{Generation time (if nucleotide substitution rates are
in days rather than generations; default: NULL, assumes that rate matrix is
per-generation)}
}
\value{
Nucletoide substitution rate matrix Q
}
\description{
Method:
\itemize{
\item If needed, convert rates per day to rates per generation
\item Convert rates per generation to probabilities per generation
\item Make diagonal such that rows of probability matrix p sum to 1
\item Convert to q matrix, assuming a mutation rate of mut_rate mutations/site/generation,
by solving for q in the equation p = exp(q*mut_rate)
}
}
\examples{
calc_q_from_rates(hiv_mut_rates, 2.4e-5, 1.2)
}
