% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tr_summary_stats.R
\name{calc_sackin_norm}
\alias{calc_sackin_norm}
\title{Calculate Sackin indexed normalized by number of tips}
\usage{
calc_sackin_norm(tr)
}
\arguments{
\item{tr}{Phylogeny}
}
\value{
Normalized Sackin index
}
\description{
Calculate Sackin indexed normalized by number of tips
}
\examples{
tr <- ape::rtree(100)
calc_sackin_norm(tr)
}
