% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tr_summary_stats.R
\name{calc_tr_dists}
\alias{calc_tr_dists}
\title{#' Calculate mean internal/external branch length ratio
#'
#' @inheritParams calc_tr_stats
#'
#' @return Mean internal branch length, mean external branch length, and mean
#'   internal/external branch length ratio
#' @export
#'
#' @examples
#' tr <- ape::rtree(100)
#' calc_int_over_ext(tr)
calc_int_over_ext <- function(tr) {
check_is_phylo(tr, "tr")
int_bl <- mean(tr$edge.length[tr$edge\link{, 2} > ape::Ntip(tr)])
ext_bl <- mean(tr$edge.length[tr$edge\link{, 2} <= ape::Ntip(tr)])
c(int_bl, ext_bl, int_bl / ext_bl)
}}
\usage{
calc_tr_dists(tr, tips = NULL)
}
\arguments{
\item{tr}{Rooted phylogeny}

\item{tips}{Tips use when calculating mean distance}
}
\value{
Mean root-to-tip distance and mean tip-to-tip distance
}
\description{
#' Calculate tree transition score for sampling times
#'
#' @inheritParams calc_tr_stats
#'
#' @return Number of transitions between timepoints on the phylogeny based on
#'   parsimony
#' @export
#'
#' @examples
#' tr <- ape::rtree(100)
#' times <- sample(3, 100, replace = TRUE)
#' names(times) <- tr$tip.label
#' calc_transition(tr, times)
calc_transition <- function(tr, timepoints) {
check_is_phylo(tr, "tr")
if (is.null(names(timepoints)) | !all(names(timepoints) \%in\% tr$tip.label)) {
stop(
"timepoints must be a vector named by tr tip labels, ",
"and must contain all tip labels."
)
}
phangorn::parsimony(
tr,
phangorn::phyDat(factor(timepoints), type = "USER")
)
}
Calculate tree distances (root-to-tip and tip-to-tip)
}
