% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tr_stats.R
\name{calc_tr_stats}
\alias{calc_tr_stats}
\title{Calculate tree summary statistics}
\usage{
calc_tr_stats(tr, timepoints, bl_thresh = 1e-08)
}
\arguments{
\item{tr}{Rooted phylogeny}

\item{timepoints}{Factor vector of time points named by tree tip label. The
levels should be ordered correctly (most often by sampling time). All tip
lables must be included in this vector. If you want to exclude certain
tips, you must drop them from the tree prior to using this function.}

\item{bl_thresh}{Branch length threshold under which branches are collapsed
to prior to calculations (default: 1e-08). This is the \code{tol} argument in
\code{\link[ape:multi2di]{ape::di2multi()}}.}
}
\value{
Tibble including tree summary statistics:
\itemize{
\item Mean leaf depth (normalized Sackin index)
\item Mean branch length
\item Mean internal branch length
\item Mean external branch length
\item Mean divergence (mean per-generation root-to-tip distance)
\item Mean diversity (mean per-generation tip-to-tip distance)
\item Divergence slope across timepoints
\item Diversity slope across timepoints
\item Mean number of lineages through time
}
}
\description{
See \code{vignette('analyze_output')} for more details.
}
\examples{
tr <- ape::rtree(100)
times <- factor(sample(3, 100, replace = TRUE), levels = 1:3)
names(times) <- tr$tip.label
calc_tr_stats(tr, times)
}
