% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tr_summary_stats.R
\name{calc_tr_summary_stats}
\alias{calc_tr_summary_stats}
\title{Calculate tree summary statistics}
\usage{
calc_tr_summary_stats(tr, timepoints)
}
\arguments{
\item{tr}{Phylogeny}

\item{timepoints}{Vector of time points named by tree tip label. All tip
lables must be included in this vector. If you want to exclude certain tips,
you must drop them from the tree prior to using this function.}
}
\value{
Tibble including 3 tree summary statistics:
\itemize{
\item Sackin index (\code{\link[treebalance:sackinI]{treebalance::sackinI()}})
\item Mean internal branch length (\code{\link[=calc_int_over_ext]{calc_int_over_ext()}})
\item Mean external branch length (\code{\link[=calc_int_over_ext]{calc_int_over_ext()}})
\item Mean internal over external branch length ratio (\code{\link[=calc_int_over_ext]{calc_int_over_ext()}})
\item Timepoint parsimony score (\code{\link[=calc_parsimony]{calc_parsimony()}})
}
}
\description{
See \code{vignette('analyze_output')} for more details.
}
\examples{
tr <- ape::rtree(100)
times <- sample(3, 100, replace = TRUE)
names(times) <- tr$tip.label
calc_tr_summary_stats(tr, times)
}
