% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_calc_nt_sub_probs_inputs}
\alias{check_calc_nt_sub_probs_inputs}
\title{Check calc_nt_sub_probs inputs}
\usage{
check_calc_nt_sub_probs_inputs(aln, tr, model, rearrangement)
}
\arguments{
\item{aln}{An alignment of class ape::DNAbin}

\item{tr}{Starting tree}

\item{model}{A string providing model (e.g. "GTR+G(4)+I")}

\item{rearrangement}{Type of tree rearrangements to perform, one of "none", "NNI", "stochastic" or "ratchet" (default: "none")}
}
\value{
error if inputs are incorrect
}
\description{
Check calc_nt_sub_probs inputs
}
