% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_define_growth_curve_inputs}
\alias{check_define_growth_curve_inputs}
\title{Check define_growth_curve}
\usage{
check_define_growth_curve_inputs(curve_type, gN, K, n0, gS, pK)
}
\arguments{
\item{curve_type}{type of growth, one of:
\code{constant_pop} (constant population size),
\code{linear} (constant growth),
\code{logistic} (logistic growth; default)}

\item{gN}{final sampling generation (default: 3000)}

\item{K}{carrying capacity (default: 2000)}

\item{n0}{starting population size (default: 1; used for linear and logistic)}

\item{gS}{seroconversion generation (default: 30; used for logistic)}

\item{pK}{proportion of carrying capacity population size at seroconversion
(default: 0.9; used for logistic)}
}
\value{
error if input parameters are incorrect
}
\description{
Check define_growth_curve
}
\examples{
check_define_growth_curve_inputs('logistic', 3000, 2000, 1, 30, 0.9)
}
