% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_define_sampling_scheme_inputs}
\alias{check_define_sampling_scheme_inputs}
\title{Check define_sampling_scheme}
\usage{
check_define_sampling_scheme_inputs(growth_curve, sampling_frequency, max_samp)
}
\arguments{
\item{growth_curve}{output from \code{define_growth_curve()}, or customized
growth curve with columns named \code{generation} and \code{active_cell_count}}

\item{sampling_frequency}{frequency in generations at which to record
sequences (and counts) (default: 300 generations)}

\item{max_samp}{maximum number of cells (and thus sequences) to sample in a
given generation (default: 20 sequences)}
}
\value{
error if inputs are incorrect
}
\description{
Check define_sampling_scheme
}
