% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_extract_founder_inputs}
\alias{check_extract_founder_inputs}
\title{Check extract founder inputs}
\usage{
check_extract_founder_inputs(aln, founder_name, start, end)
}
\arguments{
\item{aln}{Alignment}

\item{founder_name}{Name of founder sequence in the alignment}

\item{start}{Start of founder sequence (default: 1, i.e. beginning of sequence)}

\item{end}{End of founder sequence (default: NULL, i.e. end of sequence)}
}
\value{
error if wrong inputs
}
\description{
Check extract founder inputs
}
