% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_extract_seqs_inputs}
\alias{check_extract_seqs_inputs}
\title{Check extract founder inputs}
\usage{
check_extract_seqs_inputs(aln, founder_name, ref_name, start, end)
}
\arguments{
\item{aln}{Alignment}

\item{founder_name}{Name of founder sequence in the alignment}

\item{ref_name}{Optional name of reference sequence in the alignment.
This can be used as input to the \code{ref_seq} argument in \code{\link[=run_wavess]{run_wavess()}}
(default: NULL, i.e. no reference sequence is returned)}

\item{start}{Start of founder sequence in alignment
(default: 1, i.e. beginning of sequence)}

\item{end}{End of founder sequence in alignment
(default: NULL, i.e. end of sequence)}
}
\value{
error if wrong inputs
}
\description{
Check extract founder inputs
}
