% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_find_consensus_inputs}
\alias{check_find_consensus_inputs}
\title{Check find_consensus_inputs}
\usage{
check_find_consensus_inputs(aln, founder, ref, founder_aln)
}
\arguments{
\item{aln}{Alignment in \link[ape:DNAbin]{ape::DNAbin} format (e.g. read in with
\code{\link[ape:read.dna]{ape::read.dna()}} or \code{\link[ape:read.dna]{ape::read.FASTA()}}) that includes the founder sequence or
a reference sequence, plus a representative set of sequences for your genome
region of interest}

\item{founder}{Name of the founder sequence in the input alignment or in the
optional founder-specific alignment (\code{founder_aln})}

\item{ref}{Name of the reference sequence in the input alignment, required if
the founder sequence is not in \code{aln} (default: \code{NULL})}

\item{founder_aln}{Alignment including the reference and founder sequences,
required if the founder is not present in \code{aln} (default: \code{NULL})}
}
\value{
error if inputs are incorrect
}
\description{
Check find_consensus_inputs
}
