% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_generate_pop_samp_inputs}
\alias{check_generate_pop_samp_inputs}
\title{Check generate_pop_samp}
\usage{
check_generate_pop_samp_inputs(
  curve_type,
  gN,
  K,
  n0,
  g50,
  sampling_frequency,
  max_samp
)
}
\arguments{
\item{curve_type}{type of growth, one of:
\code{constant_pop} (constant population size),
\code{logistic} (logistic growth; default)}

\item{gN}{final sampling generation (default: 3000)}

\item{K}{carrying capacity (default: 2000)}

\item{n0}{starting population size (default: 1; used for logistic)}

\item{g50}{generation at which the population size reaches 50\% of K
(default: 25; used for logistic)}

\item{sampling_frequency}{frequency in generations at which to record
sequences (and counts) (default: 300 generations)}

\item{max_samp}{maximum number of cells (and thus sequences) to sample in a
given generation (default: 20 sequences)}
}
\value{
error if inputs are incorrect
}
\description{
Check generate_pop_samp
}
