% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_get_seq_pos_inputs}
\alias{check_get_seq_pos_inputs}
\title{Check get_seq_pos inputs}
\usage{
check_get_seq_pos_inputs(aln_df, col_name)
}
\arguments{
\item{aln_df}{Alignment data frame including a column with the sequence.}

\item{col_name}{The column name of the sequence of interest.}
}
\value{
error if inputs are incorrect
}
\description{
Check get_seq_pos inputs
}
