% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_is_0to1}
\alias{check_is_0to1}
\title{Check if a variable is between 0 and 1 inclusive}
\usage{
check_is_0to1(x, var_name)
}
\arguments{
\item{x}{variable to check}

\item{var_name}{name of variable}
}
\value{
error if variable is not between 0 and 1 inclusive
}
\description{
Check if a variable is between 0 and 1 inclusive
}
