% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_is_df}
\alias{check_is_df}
\title{Check if a variable is a dataframe or tibble}
\usage{
check_is_df(x, var_name)
}
\arguments{
\item{x}{variable to check}

\item{var_name}{name of variable}
}
\value{
error if variable is not numeric
}
\description{
Check if a variable is a dataframe or tibble
}
