% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_is_dnabin}
\alias{check_is_dnabin}
\title{Check if a variable is a DNAbin object}
\usage{
check_is_dnabin(x, var_name)
}
\arguments{
\item{x}{variable to check}

\item{var_name}{name of variable}
}
\value{
error if variable is not a DNAbin object
}
\description{
Check if a variable is a DNAbin object
}
