% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_is_phylo}
\alias{check_is_phylo}
\title{Check if a variable is a phylo object}
\usage{
check_is_phylo(x, var_name)

check_is_phylo(x, var_name)
}
\arguments{
\item{x}{variable to check}

\item{var_name}{name of variable}
}
\value{
error if variable is not a DNAbin object

error if variable is not a phylo object
}
\description{
Check if a variable is a phylo object

Check if a variable is a phylo object
}
