% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_is_pos}
\alias{check_is_pos}
\title{Check if a variable is positive}
\usage{
check_is_pos(x, var_name)
}
\arguments{
\item{x}{variable to check}

\item{var_name}{name of variable}
}
\value{
error if variable is not positive
}
\description{
Check if a variable is positive
}
\examples{
check_is_pos(1, 'test_var')
}
