% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_is_string}
\alias{check_is_string}
\title{Check if a variable is a string}
\usage{
check_is_string(x, var_name)
}
\arguments{
\item{x}{variable to check}

\item{var_name}{name of variable}
}
\value{
error if variable is not a string
}
\description{
Check if a variable is a string
}
