% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_name_in_alignment}
\alias{check_name_in_alignment}
\title{Check if name is in alignment}
\usage{
check_name_in_alignment(aln, name, aln_name, name_name)
}
\arguments{
\item{aln}{Alignment in DNAbin format}

\item{name}{Name of sequence to check}

\item{aln_name}{Name of alignment}

\item{name_name}{Name of name}
}
\value{
error if name isn't in the alignment
}
\description{
Check if name is in alignment
}
