% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_seq}
\alias{check_seq}
\title{Check check_seq sequence input}
\usage{
check_seq(seq, chars, seq_type)
}
\arguments{
\item{seq}{character sequence}

\item{chars}{characters}

\item{seq_type}{type of sequence}
}
\value{
error if wrong characters in sequence
}
\description{
Check check_seq sequence input
}
