% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_slice_aln_inputs}
\alias{check_slice_aln_inputs}
\title{Check slice_aln inputs}
\usage{
check_slice_aln_inputs(aln, start, end, seqs)
}
\arguments{
\item{aln}{alignment to slice}

\item{start}{start position in alignment}

\item{end}{end position in alignment}

\item{seqs}{sequences to keep (default: labels(aln), i.e. all sequences)}
}
\value{
error if wrong inputs
}
\description{
Check slice_aln inputs
}
