% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_epitopes.R
\name{convert_ref_to_founder_epitopes}
\alias{convert_ref_to_founder_epitopes}
\title{Convert reference epitope locations to founder epitope locations}
\usage{
convert_ref_to_founder_epitopes(ref_epitopes, ref_founder_map)
}
\arguments{
\item{ref_epitopes}{Output from \code{\link[=sample_epitopes]{sample_epitopes()}}, including start and end
nt positions of the epitopes (\code{epi_start_nt} and \code{epi_end_nt})}

\item{ref_founder_map}{Output from \code{\link[=map_ref_founder]{map_ref_founder()}}, including reference
and founder positions (\code{ref_pos} and \code{founder_pos}).}
}
\value{
Tibble with epitope positions relative to the founder, with the
same columns as output by \code{\link[=sample_epitopes]{sample_epitopes()}}
}
\description{
Convert reference epitope locations to founder epitope locations
}
