% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pop_samp.R
\name{define_growth_curve}
\alias{define_growth_curve}
\title{Define growth curve}
\usage{
define_growth_curve(
  curve_type = "logistic",
  gN = 3000,
  K = 2000,
  n0 = 1,
  gpK = 25,
  pK = 0.5
)
}
\arguments{
\item{curve_type}{type of growth, one of:
\code{constant_pop} (constant population size),
\code{logistic} (logistic growth; default)}

\item{gN}{final sampling generation (default: 3000)}

\item{K}{carrying capacity (default: 2000)}

\item{n0}{starting population size (default: 1; used for logistic)}

\item{gpK}{generation at which the population size reaches pK of K
(default: 25; used for logistic)}

\item{pK}{proportion of carrying capacity population size at gpK
(default: 0.5; used for logistic)}
}
\value{
tibble with two columns: generation and active cell count
}
\description{
Define growth curve
}
