% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_sampling_scheme.R
\name{define_sampling_scheme}
\alias{define_sampling_scheme}
\title{Define sampling scheme}
\usage{
define_sampling_scheme(growth_curve, sampling_frequency = 300, max_samp = 20)
}
\arguments{
\item{growth_curve}{output from \code{define_growth_curve()}, or customized
growth curve with columns named \code{generation} and \code{active_cell_count}}

\item{sampling_frequency}{frequency in generations at which to record
sequences (and counts) (default: 300 generations)}

\item{max_samp}{maximum number of cells (and thus sequences) to sample in a
given generation (default: 20 sequences)}
}
\value{
input growth curve tibble with one additional column (\code{n_sample_active})
containing the number of sequences from active cells to samples
}
\description{
Define sampling scheme
}
\examples{
define_growth_curve() |> define_sampling_scheme()
}
