% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{env_features}
\alias{env_features}
\title{ENV features}
\format{
\subsection{\code{env_features}}{

A tibble of HIV env binding, contact, and neutralization sites curated by the
LANL HIV database. The full Env feature dataset was subsetted to include only
binding, contact, and neutralization sites. The most relevant column for us
is the "position" column. Other columns are included in case it is useful for
specific use cases. This particular data was last updated on 2024-10-01.
}
}
\source{
\url{https://www.hiv.lanl.gov/components/sequence/HIV/neutralization/download_db.comp}
}
\usage{
env_features
}
\description{
Binding, contact, and neutralization features in HIV env gene
}
\keyword{datasets}
