% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_q.R
\name{estimate_q}
\alias{estimate_q}
\title{Calculate nucleotide substitution probabilities}
\usage{
estimate_q(aln, tr = NULL, model = "GTR+R(4)+I", rearrangement = "none")
}
\arguments{
\item{aln}{An alignment of class ape::DNAbin}

\item{tr}{Starting tree}

\item{model}{A string providing model (e.g. "GTR+G(4)+I")}

\item{rearrangement}{Type of tree rearrangements to perform, one of
"none", "NNI", "stochastic" or "ratchet" (default: "none")}
}
\value{
Matrix of nucleotide substitution probabilities.
Columns are from and rows are to.
}
\description{
For this, you should probably use a set of within-host sequences for the
genome region of interest.
}
\examples{
# NOTE: This is just an example.
estimate_q(hiv_env_flt_2022)
}
