% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_founder.R
\name{extract_founder}
\alias{extract_founder}
\title{Extract founder sequence from an alignment}
\usage{
extract_founder(aln, founder_name, start = 1, end = NULL)
}
\arguments{
\item{aln}{Alignment}

\item{founder_name}{Name of founder sequence in the alignment}

\item{start}{Start of founder sequence (default: 1, i.e. beginning of sequence)}

\item{end}{End of founder sequence (default: NULL, i.e. end of sequence)}
}
\value{
Founder sequence as a character string
}
\description{
Extract founder sequence from an alignment
}
\examples{
hxb2_founder <- ape::as.matrix.DNAbin(hxb2_founder)
extract_founder(hxb2_founder, 'B.US.2011.DEMB11US006.KC473833')
}
