% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_seqs.R
\name{extract_seqs}
\alias{extract_seqs}
\title{Extract founder (and reference) sequence from an alignment}
\usage{
extract_seqs(aln, founder_name, ref_name = NULL, start = 1, end = NULL)
}
\arguments{
\item{aln}{Alignment}

\item{founder_name}{Name of founder sequence in the alignment}

\item{ref_name}{Optional name of reference sequence in the alignment.
This can be used as input to the \code{ref_seq} argument in \code{\link[=run_wavess]{run_wavess()}}
(default: NULL, i.e. no reference sequence is returned)}

\item{start}{Start of founder sequence in alignment
(default: 1, i.e. beginning of sequence)}

\item{end}{End of founder sequence in alignment
(default: NULL, i.e. end of sequence)}
}
\value{
List of founder sequence and optional reference sequence
as character strings
}
\description{
Extract founder (and reference) sequence from an alignment
}
\examples{
extract_seqs(hxb2_cons_founder, "B.US.2011.DEMB11US006.KC473833", start = 6225, end = 7757)
extract_seqs(hxb2_cons_founder,
  "B.US.2011.DEMB11US006.KC473833", "B.FR.83.HXB2_LAI_IIIB_BRU.K03455",
  start = 6225, end = 7757
)
}
