% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_conserved_sites.R
\name{find_consensus}
\alias{find_consensus}
\title{Determine consensus sequence}
\usage{
find_consensus(aln, founder, ref = NULL, founder_aln = NULL)
}
\arguments{
\item{aln}{Alignment in \code{ape::DNAbin} format (e.g. read in with
\code{ape::read.dna()} or \code{ape::read.FASTA()}) that includes the founder sequence or
a reference sequence, plus a representative set of sequences for your genome
region of interest}

\item{founder}{Name of the founder sequence in the input alignment or in the
optional founder-specific alignment (\code{founder_aln})}

\item{ref}{Name of the reference sequence in the input alignment, required if
the founder sequence is not in \code{aln} (default: \code{NULL})}

\item{founder_aln}{Alignment including the reference and founder sequences,
required if the founder is not present in \code{aln} (default: \code{NULL})}
}
\value{
Tibble including the following columns:
\itemize{
\item \code{founder_pos}: founder position
\item \code{founder_base}: founder base
\item \code{consensus_base}: consensus base
\item \code{consensus_prop}: proportion of sequences that had that base at that position
When using a reference, \code{NA} in the consensus columns indicates that that
position was an insertion relative to the reference
}
}
\description{
This function takes a very simple approach to determining the consensus
sequence. It simply finds the majority base. If there is a tie, it takes the
base that comes first alphabetically. For more nuanced consensus-making,
please check out the
\href{https://www.hiv.lanl.gov/content/sequence/CONSENSUS/consensus.html}{LANL HIV database consensus maker}
}
