% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{founder_conserved_sites}
\alias{founder_conserved_sites}
\title{Example founder conserved sites}
\format{
\subsection{\code{founder_conserved_sites} Vector of conserved nucleotides for the}{

gp120 gene of DEMB11US006 (founder), named by the sequence position (indexed at 0)
}
}
\source{
Full alignment used:
\url{https://www.hiv.lanl.gov/content/sequence/NEWALIGN/align.html}
Founder sequence:
\url{https://www.sciencedirect.com/science/article/pii/S0022175914000143?via\%3Dihub}
\url{https://www.hiv.lanl.gov/components/sequence/HIV/asearch/query_one.comp?se_id=KC473833}
}
\usage{
founder_conserved_sites
}
\description{
Conserved sites (indexed at 1) for the gp120 DEMB11US006 sequence, which is
used as an example founder sequence throughout the package. These sites were
identified using the \code{identify_conserved_sites()} function with the full
HIV ENV filtered alignment from the LANL HIV database.
See \code{hiv_env_flt_2022} for more details and the link to the entire alignment.
}
\keyword{datasets}
