% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pop_samp.R
\name{generate_pop_samp}
\alias{generate_pop_samp}
\title{Generate population growth curve and sampling scheme}
\usage{
generate_pop_samp(
  curve_type = "logistic",
  gN = 3000,
  K = 2000,
  n0 = 1,
  gS = 30,
  pK = 0.9,
  sampling_frequency = 300,
  max_samp = 20
)
}
\arguments{
\item{curve_type}{type of growth, one of:
\code{constant_pop} (constant population size),
\code{logistic} (logistic growth; default)}

\item{gN}{final sampling generation (default: 3000)}

\item{K}{carrying capacity (default: 2000)}

\item{n0}{starting population size (default: 1; used for logistic)}

\item{gS}{seroconversion generation (default: 30; used for logistic)}

\item{pK}{proportion of carrying capacity population size at seroconversion
(default: 0.9; used for logistic)}

\item{sampling_frequency}{frequency in generations at which to record
sequences (and counts) (default: 300 generations)}

\item{max_samp}{maximum number of cells (and thus sequences) to sample in a
given generation (default: 20 sequences)}
}
\value{
tibble with two columns:
\itemize{
\item \code{generation}: Each generation to be simulated
\item \code{active_cell_count}: Number of active cells in each generation
\item \code{n_sample_active} Number of sequences to sample in each generation
}
}
\description{
Generate population growth curve and sampling scheme
}
\examples{
generate_pop_samp()
}
