% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tr_summary_stats.R
\name{get_clusters}
\alias{get_clusters}
\title{Get monophyletic clusters on the phylogeny}
\usage{
get_clusters(tr, timepoints, pureness = 1, bootstrap = NULL, grps = NULL)
}
\arguments{
\item{tr}{Phylogeny}

\item{timepoints}{Vector of time points named by tree tip label}

\item{pureness}{How pure each cluster should be (must be > 0.5) (optional, default = 1)}

\item{bootstrap}{Bootstrap support to use to filter unconfident tree edges (optional, default = NULL)}

\item{grps}{Groups from which to keep only one isolate (default: NULL, all kept)}
}
\value{
list where pure_subtree_info is a data.frame of clusters on phylogeny,
index indicates which element that cluster is in the list of subtrees,
NA indicates it is not part of a subtree;
subtrees is an object of the actual subtrees (can be used for plotting);
cluster_pureness is the purness of each cluster
}
\description{
This code was modified from https://github.com/Snitkin-Lab-Umich/regentrans/,
which is under an MIT license
}
