% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_epitopes.R
\name{get_epitope_frequencies}
\alias{get_epitope_frequencies}
\title{Get epitope frequencies}
\usage{
get_epitope_frequencies(epitope_positions)
}
\arguments{
\item{epitope_positions}{numeric vector of \strong{amino acid} positions of
identified epitopes (e.g. the env features from the \href{https://www.hiv.lanl.gov/components/sequence/HIV/neutralization/download_db.comp}{LANL HIV database})}
}
\value{
tibble with the following columns:
\itemize{
\item \code{aa_position}: amino acid position
\item \code{n_features}: number of features at that position
\item \code{epitope_probability}: estimated probability of an epitope at that position
}
}
\description{
Get epitope frequencies
}
\examples{
get_epitope_frequencies(env_features$Position-1) # subtract 1 to index at 0
}
