% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_growth_curve.R
\name{get_logistic_curve}
\alias{get_logistic_curve}
\title{Get logistic growth curve}
\usage{
get_logistic_curve(n0, K, gS, pK, gN)
}
\arguments{
\item{n0}{starting population size (default: 1; used for linear and logistic)}

\item{K}{carrying capacity (default: 2000)}

\item{gS}{seroconversion generation (default: 30; used for logistic)}

\item{pK}{proportion of carrying capacity population size at seroconversion
(default: 0.9; used for logistic)}

\item{gN}{final sampling generation (default: 3000)}
}
\value{
Population size at each generation
}
\description{
Note: to get the equations used in this function, I solved for k and xM using
}
\examples{
get_logistic_curve(1, 2000, 30, 0.9, 3000)
}
