% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_conserved_sites.R
\name{get_seq_pos}
\alias{get_seq_pos}
\title{Get sequence position}
\usage{
get_seq_pos(aln_df, col_name)
}
\arguments{
\item{aln_df}{Alignment data frame including a column with the sequence.}

\item{col_name}{The column name of the sequence of interest.}
}
\value{
Original data frame with an additional column (the column name prepended to
_pos) that indicates the position of that base in the sequence.n
}
\description{
Get sequence position from alignment data frame.
Currently considers everything a position that isn't '-'
}
