% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hiv_env_flt_2021}
\alias{hiv_env_flt_2021}
\title{HIV1 ENV alignment}
\format{
\subsection{\code{hiv_env_flt_2021}}{

An \code{ape} DNAbin object with 6,741 DNA sequences of length 3993
}
}
\source{
\url{https://www.hiv.lanl.gov/content/sequence/NEWALIGN/align.html}
}
\usage{
hiv_env_flt_2021
}
\description{
The 2021 HIV1 ENV FLT web alignment from the LANL HIV database
}
\keyword{datasets}
