% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hiv_env_flt_2022}
\alias{hiv_env_flt_2022}
\title{HIV1 ENV alignment}
\format{
\subsection{\code{hiv_env_flt_2022}}{

An \code{ape} DNAbin object with 10 DNA sequences of length 3993.
If you would like to download the entire set of 6,741 sequences, please
visit the link below.
}
}
\source{
\url{https://www.hiv.lanl.gov/content/sequence/NEWALIGN/align.html}
}
\usage{
hiv_env_flt_2022
}
\description{
The 2022 HIV1 ENV FLT web alignment from the LANL HIV database
}
\keyword{datasets}
