% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hiv_q_mat}
\alias{hiv_q_mat}
\title{HIV Q matrix}
\format{
\subsection{\code{hiv_q_mat}}{

A matrix containing relative rates of change between nucleotides, where the
rows are the "from" nucleotide and the columns are the "to" nucleotide.
To compute this we used:
\itemize{
\item Individual substitution rates from Zanini et al. 2017
\item Base frequencies from the LANL 2022 HIV ENV filtered alignment
\item The overall substitution rate from Mansky 1996.
}
}
}
\source{
Zanini et al. 2017: \url{https://doi.org/10.1093/ve/vex003}
HIV ENV filtered alignment: \url{https://www.hiv.lanl.gov/content/sequence/NEWALIGN/align.html}
Mansky 1996: \url{https://doi.org/10.1089/aid.1996.12.307}
}
\usage{
hiv_q_mat
}
\description{
Q matrix containing relative rates of change between nucleotides.
}
\keyword{datasets}
