% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hxb2_founder}
\alias{hxb2_founder}
\title{gp120 HXB2 and founder alignment}
\format{
\subsection{\code{hxb2_founder}}{

An \code{ape} DNAbin object with 2 DNA sequences of length 1563 including the
env gp120 sequence for hxb2 and DEMB11US006 (founder)
}
}
\source{
HXB2 (1st sequence):
\url{https://www.hiv.lanl.gov/content/sequence/NEWALIGN/align.html}
Founder:
\url{https://www.sciencedirect.com/science/article/pii/S0022175914000143?via\%3Dihub}
\url{https://www.hiv.lanl.gov/components/sequence/HIV/asearch/query_one.comp?se_id=KC473833}
}
\usage{
hxb2_founder
}
\description{
Alignment of gp120 HXB2 and DEMB11US006 (founder)
}
\keyword{datasets}
