% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_conserved_sites.R
\name{identify_conserved_sites}
\alias{identify_conserved_sites}
\title{Identify conserved sites}
\usage{
identify_conserved_sites(
  aln,
  founder,
  thresh = 0.99,
  ref = NULL,
  founder_aln = NULL
)
}
\arguments{
\item{aln}{Alignment in \link[ape:DNAbin]{ape::DNAbin} format (e.g. read in with
\code{\link[ape:read.dna]{ape::read.dna()}} or \code{\link[ape:read.FASTA]{ape::read.FASTA()}}) that includes the founder
sequence or a reference sequence, plus a representative set of sequences
for your genome region of interest}

\item{founder}{Name of the founder sequence in the input alignment or in the
optional founder-specific alignment (\code{founder_aln})}

\item{thresh}{Conserved site threshold. A position is considered to be
conserved if >thresh proportion of sequences in the alignment are the same
base (default: 0.99)}

\item{ref}{Name of the reference sequence in the input alignment, required if
the founder sequence is not in \code{aln} (default: \code{NULL})}

\item{founder_aln}{Alignment including the reference and founder sequences,
required if the founder is not present in \code{aln}. \strong{NOTE:} This alignment
and \code{aln} are assumed to begin at the same position in the reference
sequence (default: \code{NULL})}
}
\value{
Tibble including the following columns:
\itemize{
\item \code{founder_pos}: founder position
\item \code{founder_base}: founder base
\item \code{consensus_base}: consensus base
\item \code{consensus_prop}: proportion of sequences that had that base at that
position
\item \code{conserved}: whether or not the position is conserved (Yes means conserved,
No means not conserved, NA means the conserved position is a gap ('-'))
When using a reference, \code{NA} in the consensus columns indicates that that
position was an insertion relative to the reference. All positions are
indexed at 0.
}
}
\description{
Identify conserved sites relative to the founder sequence. Note that the
positions returned assume that the start of the alignment is also the start
of the founder sequence in the simulation.
}
\examples{
gp120_flt_2022 <- slice_aln(hiv_env_flt_2022, start = 1, end = 2517)
gp120_hxb2_cons_founder <- slice_aln(hxb2_cons_founder, start = 6225, end = 7757)
identify_conserved_sites(
  gp120_flt_2022,
  "B.FR.83.HXB2_LAI_IIIB_BRU.K03455"
)
identify_conserved_sites(gp120_flt_2022,
  "B.US.2011.DEMB11US006.KC473833",
  ref = "B.FR.83.HXB2_LAI_IIIB_BRU.K03455",
  founder_aln = gp120_hxb2_cons_founder
)
}
