% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_growth_curve.R
\name{logistic_fn}
\alias{logistic_fn}
\title{Get y value of logistic curve}
\usage{
logistic_fn(x, K, growth_rate, midpoint)
}
\arguments{
\item{x}{x value of logistic function (generation)}

\item{K}{carrying capacity (default: 2000)}

\item{growth_rate}{growth rate}

\item{midpoint}{midpoint of logistic curve}
}
\value{
y value of logistic curve
}
\description{
Get the y value for a logistic function given the
x value, carrying capacity, growth rate, and midpoint
}
