% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_conserved_sites.R
\name{map_ref_founder}
\alias{map_ref_founder}
\title{Map reference and founder sequence positions}
\usage{
map_ref_founder(aln, ref, founder)
}
\arguments{
\item{aln}{Alignment in \link[ape:DNAbin]{ape::DNAbin} format (e.g. read in with
\code{\link[ape:read.dna]{ape::read.dna()}} or \code{\link[ape:read.FASTA]{ape::read.FASTA()}}) that includes the reference
sequence and the founder sequence}

\item{ref}{Name of the reference sequence in the input alignment}

\item{founder}{Name of the founder sequence in the input alignment}
}
\value{
Tibble including columns for:
\itemize{
\item Alignment position (\code{alignment_pos})
\item Reference base position (\code{ref_pos})
\item Founder base position (\code{founder_pos})
\item Reference base (\code{ref_base})
\item Founder base (\code{founder_base})
All positions are indexed at 0.
}
}
\description{
Map reference and founder sequence positions
}
\examples{
map_ref_founder(
  hxb2_cons_founder,
  "B.FR.83.HXB2_LAI_IIIB_BRU.K03455",
  "B.US.2011.DEMB11US006.KC473833"
)
}
