% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_python_dependencies.R
\name{py_check_installed}
\alias{py_check_installed}
\title{Check if python dependency is installed}
\usage{
py_check_installed(x)
}
\arguments{
\item{x}{dependency to check}
}
\value{
boolean whether dependency is installed
}
\description{
Check if python dependency is installed
}
