% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_python_dependencies.R
\name{py_check_version}
\alias{py_check_version}
\title{Check python package version}
\usage{
py_check_version(package, ...)
}
\arguments{
\item{package}{Package to check version of}

\item{...}{Extra stuff for reticulate::py_list_packages}
}
\value{
package version
}
\description{
Check python package version
}
