% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_to_probability.R
\name{rate_to_probability}
\alias{rate_to_probability}
\title{Convert a rate to a Poisson probability}
\usage{
rate_to_probability(rate, k = 1)
}
\arguments{
\item{rate}{Rate to be converted}

\item{k}{Number of events observed in the rate-defined time period (default: 1)}
}
\value{
Probability that the event occurs in the rate-defined time period
}
\description{
The \code{wavess()} function takes probabilities,
but sometimes in the literature you find rates instead.
This funciton allows you to convert that rate to a probability.
}
\examples{
rate_to_probability(2.1)
rate_to_probability(0.1)
rate_to_probability(3.5e-5)
}
