% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_wavess.R
\name{record_counts}
\alias{record_counts}
\title{Record counts}
\usage{
record_counts(counts, generation, host, latent_nums, var_nums, fitness)
}
\arguments{
\item{counts}{Empty tibble or previous counts}

\item{generation}{Generation}

\item{host}{Host environment (python)}

\item{latent_nums}{Numbers related to latency (num_to_make_latent, num_to_activate, num_to_die, num_to_proliferate)}

\item{var_nums}{Numbers related to variation (n_mut, number_recombination)}

\item{fitness}{Numbers related to fitness (mean_fitness_active, mean_conserved_cost_active, mean_immune_cost_active, mean_replicative_cost_active)}
}
\value{
Updated counts tibble with additional row added for generation
}
\description{
Record counts
}
