% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_python_dependencies.R
\name{remove_python_venv}
\alias{remove_python_venv}
\title{Uninstall the wavess environment}
\usage{
remove_python_venv(confirm = interactive())
}
\arguments{
\item{confirm}{logical; confirm before uninstalling wavess virtual environment?}
}
\description{
Removes the virtual environment created by create_python_venv()
}
