% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_epitopes.R
\name{sample_epitopes}
\alias{sample_epitopes}
\title{Sample epitopes}
\usage{
sample_epitopes(
  epitope_probabilities,
  start_aa_pos = 1,
  end_aa_pos = NULL,
  num_epitopes = 10,
  aa_epitope_length = 10,
  max_fit_cost = 0.4,
  cost_type = "linear",
  max_resamples = 100,
  ref_founder_map = NULL
)
}
\arguments{
\item{epitope_probabilities}{Epitope probability tibble as output by
\code{get_epitope_frequencies()}, including columns
\code{aa_position} and \code{epitope_probability}}

\item{start_aa_pos}{Starting amino acid position to consider for epitopes
(default: 1)}

\item{end_aa_pos}{Ending amino acid position to consider for epitopes
(default: NULL, i.e. all positions)}

\item{num_epitopes}{Number of epitopes to sample}

\item{aa_epitope_length}{Amino acid epitope length}

\item{max_fit_cost}{Maximum fitness cost of the epitope,
must be between 0 and 1 (default: 0.4)
\strong{note that the model output is very sensitive to this parameter}}

\item{cost_type}{"linear" or "random"; linear returns max fitness costs
distributed linearly between 0 and \code{max_fit_cost}, random
returns max fitness costs randomly selected from a uniform distribution between
0 and \code{max_fit_cost} (default: "linear")}

\item{max_resamples}{Maximum number of resampling events to attempt;
this is to prevent an infinite loop (default: 100)}

\item{ref_founder_map}{Output from \code{map_ref_founder()}, including reference
and founder positions (\code{ref_pos} and \code{founder_pos}).}
}
\value{
tibble with the \code{num_epitopes} rows and the following columns:
\itemize{
\item \code{epi_start_nt}: nucleotide epitope start position (indexed at 0)
\item \code{epi_end_nt}: nucleotide epitope end position (indexed at 0)
\item \code{max_fitness_cost}: maximum fitness cost for that epitope
}
}
\description{
Sample epitopes
}
\examples{
sample_epitopes(get_epitope_frequencies(env_features$position))
}
