% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_wavess.R
\name{sample_viral_sequences}
\alias{sample_viral_sequences}
\title{Sample viral sequences}
\usage{
sample_viral_sequences(generation, host, n_to_samp)
}
\arguments{
\item{generation}{Generation (to include in sequence name)}

\item{host}{Host environment (python)}

\item{n_to_samp}{Number of cells to sample}
}
\value{
Sampled sequences in \link[ape:DNAbin]{ape::DNAbin} format.
}
\description{
Sample viral sequences
}
