% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_seqs.R
\name{slice_aln}
\alias{slice_aln}
\title{Slice alignment}
\usage{
slice_aln(aln, start, end, seqs = labels(aln))
}
\arguments{
\item{aln}{alignment to slice}

\item{start}{start position in alignment}

\item{end}{end position in alignment}

\item{seqs}{sequences to keep (default: labels(aln), i.e. all sequences)}
}
\value{
sliced alignment in \code{ape::DNAbin} format
}
\description{
Slice alignment
}
\examples{
slice_aln(hxb2_cons_founder, 1, 100)
}
