% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tr_summary_stats.R
\name{calc_tr_dists}
\alias{calc_tr_dists}
\title{Calculate tree distances (root-to-tip and tip-to-tip)}
\usage{
calc_tr_dists(tr)
}
\arguments{
\item{tr}{Rooted phylogeny}
}
\value{
Mean root-to-tip distance and mean tip-to-tip distance
}
\description{
Calculate tree distances (root-to-tip and tip-to-tip)
}
\examples{
tr <- ape::rtree(100)
calc_tr_dists(tr)
}
