% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tr_stats.R
\name{calc_tr_stats}
\alias{calc_tr_stats}
\title{Calculate tree summary statistics}
\usage{
calc_tr_stats(tr, timepoints)
}
\arguments{
\item{tr}{Rooted phylogeny}

\item{timepoints}{Vector of time points named by tree tip label. All tip
lables must be included in this vector. If you want to exclude certain tips,
you must drop them from the tree prior to using this function.}
}
\value{
Tibble including 3 tree summary statistics:
\itemize{
\item Mean leaf depth (normalized Sackin index)
\item Mean internal branch length
\item Mean external branch length
\item Timepoint transition score
\item Mean tip-to-tip distance
\item Mean divergence (per-generation root-to-tip distance)
\item Mean diversity (per-generation tip-to-tip distance)
}
}
\description{
See \code{vignette('analyze_output')} for more details.
}
\examples{
tr <- ape::rtree(100)
times <- sample(3, 100, replace = TRUE)
names(times) <- tr$tip.label
calc_tr_stats(tr, times)
}
