% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tr_summary_stats.R
\name{calc_tr_summary_stats}
\alias{calc_tr_summary_stats}
\title{Calculate tree summary statistics}
\usage{
calc_tr_summary_stats(tr, timepoints)
}
\arguments{
\item{tr}{Phylogeny}

\item{timepoints}{Vector of time points named by tree tip label}
}
\value{
Tibble including 3 tree summary statistics:
\itemize{
\item Sackin index normalized by the number of tree tips
\item Mean internal over external branch length ratio
\item Mean percent of lineages that survived from one time point to the next
}
}
\description{
Calculate tree summary statistics
}
\examples{
tr <- ape::rtree(100)
times <- sample(3, 100, replace = TRUE)
names(times) <- tr$tip.label
calc_tr_summary_stats(tr, times)
}
