% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_identify_conserved_sites_inputs}
\alias{check_identify_conserved_sites_inputs}
\title{Check identify_conserved_sites inputs}
\usage{
check_identify_conserved_sites_inputs(
  aln,
  founder,
  thresh,
  ref,
  founder_aln,
  founder_start_pos
)
}
\arguments{
\item{aln}{Alignment in \code{ape::DNAbin} format (e.g. read in with
\code{ape::read.dna()} or \code{ape::read.FASTA()}) that includes the founder sequence or
a reference sequence, plus a representative set of sequences for your genome
region of interest}

\item{founder}{Name of the founder sequence in the input alignment or in the
optional founder-specific alignment (\code{founder_aln})}

\item{thresh}{Conserved site threshold. A position is considered to be
conserved if >thresh proportion of sequences in the alignment are the same
base (default: 0.99)}

\item{ref}{Name of the reference sequence in the input alignment, required if
the founder sequence is not in \code{aln} (default: \code{NULL})}

\item{founder_aln}{Alignment including the reference and founder sequences,
required if the founder is not present in \code{aln} (default: \code{NULL})}

\item{founder_start_pos}{Starting position of the founder sequence in \code{aln}
(default: 1). This start position is used to change the indexing of the
positions in the original alignment so the first base in the
founder is 0, and (if needed) to match \code{aln} and \code{founder_aln} based on the
chosen reference. NOTE: this function assumes that the founder start position
of \code{founder_aln} is 1.}
}
\value{
error if inputs are incorrect
}
\description{
Check identify_conserved_sites inputs
}
