% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_is_pos}
\alias{check_is_pos}
\title{Check if a variable is positive}
\usage{
check_is_pos(x, var_name, ok0 = FALSE)
}
\arguments{
\item{x}{variable to check}

\item{var_name}{name of variable}

\item{ok0}{whether 0 is okay (default: FALSE)}
}
\value{
error if variable is not positive
}
\description{
Check if a variable is positive
}
