% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_map_ref_founder_inputs}
\alias{check_map_ref_founder_inputs}
\title{Check map_ref_founder inputs}
\usage{
check_map_ref_founder_inputs(aln, ref, founder)
}
\arguments{
\item{aln}{Alignment in \link[ape:DNAbin]{ape::DNAbin} format (e.g. read in with
\code{\link[ape:read.dna]{ape::read.dna()}} or \code{\link[ape:read.dna]{ape::read.FASTA()}}) that includes the reference sequence
and the founder sequence}

\item{ref}{Name of the reference sequence in the input alignment}

\item{founder}{Name of the founder sequence in the input alignment}
}
\value{
error if inputs are incorrect
}
\description{
Check map_ref_founder inputs
}
