% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_growth_curve.R
\name{define_growth_curve}
\alias{define_growth_curve}
\title{Define growth curve}
\usage{
define_growth_curve(
  curve_type = "logistic",
  gN = 3000,
  K = 2000,
  n0 = 1,
  gS = 30,
  pK = 0.9
)
}
\arguments{
\item{curve_type}{type of growth, one of:
\code{constant_pop} (constant population size),
\code{linear} (constant growth),
\code{logistic} (logistic growth; default)}

\item{gN}{final sampling generation (default: 3000)}

\item{K}{carrying capacity (default: 2000)}

\item{n0}{starting population size (default: 1; used for linear and logistic)}

\item{gS}{seroconversion generation (default: 30; used for logistic)}

\item{pK}{proportion of carrying capacity population size at seroconversion
(default: 0.9; used for logistic)}
}
\value{
tibble with two columns: generation and active cell count
}
\description{
Define growth curve
}
\examples{
define_growth_curve()
define_growth_curve(curve_type = 'constant')
define_growth_curve(gN = 1000)
}
