% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pop_samp.R
\name{define_sampling_scheme}
\alias{define_sampling_scheme}
\title{Define sampling scheme}
\usage{
define_sampling_scheme(sampling_frequency = 365, max_samp = 20, n_days = 3650)
}
\arguments{
\item{sampling_frequency}{frequency in days at which to record
sequences (and counts) (default: 365 days)}

\item{max_samp}{maximum number of cells (and thus sequences) to sample in a
given day (default: 20 sequences)}

\item{n_days}{day to end sampling (default: 3650)}
}
\value{
input growth curve tibble with one additional column
(\code{n_sample_active}) containing the number of sequences from active cells to
samples
}
\description{
Define which days to sample sequences, and how many sequences to sample
}
\examples{
define_growth_curve()
}
