% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pop_samp.R
\name{generate_pop_samp}
\alias{generate_pop_samp}
\title{Generate population growth curve and sampling scheme}
\usage{
generate_pop_samp(
  n_gen = 3000,
  carry_cap = 2000,
  n0 = 1,
  g50 = 25,
  sampling_frequency = 300,
  max_samp = 20
)
}
\arguments{
\item{n_gen}{final sampling generation (default: 3000)}

\item{carry_cap}{carrying capacity (default: 2000)}

\item{n0}{starting population size (default: 1; used for logistic)}

\item{g50}{generation at which the population size reaches 50\% of carry_cap
(default: 25; used for logistic)}

\item{sampling_frequency}{frequency in generations at which to record
sequences (and counts) (default: 300 generations)}

\item{max_samp}{maximum number of cells (and thus sequences) to sample in a
given generation (default: 20 sequences)}
}
\value{
tibble with two columns:
\itemize{
\item \code{generation}: Each generation to be simulated
\item \code{active_cell_count}: Number of active cells in each generation
\item \code{n_sample_active} Number of sequences to sample in each generation
}
}
\description{
Generate population growth curve and sampling scheme
}
\examples{
generate_pop_samp()
}
