% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pop_samp.R
\name{get_logistic_curve}
\alias{get_logistic_curve}
\title{Get logistic growth curve}
\usage{
get_logistic_curve(n0, K, gS, pK, gN)
}
\arguments{
\item{n0}{starting population size (default: 1; used for logistic)}

\item{K}{carrying capacity (default: 2000)}

\item{pK}{proportion of carrying capacity population size at gpK
(default: 0.5; used for logistic)}

\item{gN}{final sampling generation (default: 3000)}
}
\value{
Population size at each generation
}
\description{
Note: to get the equations used in this function, I solved for the
growth rate and midpoint based on n0, K, gS, and pK.
}
