% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hiv_q_mat}
\alias{hiv_q_mat}
\title{HIV Q matrix}
\format{
\subsection{\code{hiv_q_mat}}{

A matrix containing relative rates of change between nucleotides, where the
rows are the "from" nucleotide and the columns are the "to" nucleotide.
To compute this we used substitution rates from Zanini et al. 2017
}
}
\source{
Zanini et al. 2017: \url{https://doi.org/10.1093/ve/vex003}
}
\usage{
hiv_q_mat
}
\description{
Q matrix containing relative rates of change between nucleotides.
}
\keyword{datasets}
