% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hiv_q_mat}
\alias{hiv_q_mat}
\title{HIV Q matrix estimated from approximately neutral sites}
\format{
\subsection{\code{hiv_q_mat}}{

A Q matrix containing relative rates of change between nucleotides, where the
rows are the "from" nucleotide and the columns are the "to" nucleotide.
}
}
\source{
Zanini et al. 2017: \url{https://doi.org/10.1093/ve/vex003}
Mansky 1996: \url{https://doi.org/10.1089/aid.1996.12.307}
}
\usage{
hiv_q_mat
}
\description{
Q matrix for HIV containing relative rates of change between
nucleotides, from the nucleotide substitution rates estimated in
Zanini et al. 2017.
Method for computing Q matrix:
\itemize{
\item convert rates per day to rates per generation
\item convert rates per generation to probabilities per generation
\item make diagonal such that rows of probability matrix p sum to 1
\item convert to q matrix, assuming a substitution rate t of 3.5e-5
substitutions/site/generation (Mansky 1996), by solving for q in the
equation p = exp(qt)
}
}
\keyword{datasets}
