% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_conserved_sites.R
\name{identify_conserved_sites}
\alias{identify_conserved_sites}
\title{Identify conserved sites}
\usage{
identify_conserved_sites(
  aln,
  founder,
  thresh = 0.99,
  ref = NULL,
  founder_aln = NULL,
  founder_start_pos = 1
)
}
\arguments{
\item{aln}{Alignment in \code{ape::DNAbin} format (e.g. read in with
\code{ape::read.dna()} or \code{ape::read.FASTA()}) that includes the founder sequence or
a reference sequence, plus a representative set of sequences for your genome
region of interest}

\item{founder}{Name of the founder sequence in the input alignment or in the
optional founder-specific alignment (\code{founder_aln})}

\item{thresh}{Conserved site threshold. A position is considered to be
conserved if >thresh proportion of sequences in the alignment are the same
base (default: 0.99)}

\item{ref}{Name of the reference sequence in the input alignment, required if
the founder sequence is not in \code{aln} (default: \code{NULL})}

\item{founder_aln}{Alignment including the reference and founder sequences,
required if the founder is not present in \code{aln} (default: \code{NULL})}

\item{founder_start_pos}{Starting position of the founder sequence in \code{aln}
(default: 1). This start position is used to change the indexing of the
positions in the original alignment so the first base in the
founder is 0, and (if needed) to match \code{aln} and \code{founder_aln} based on the
chosen reference. NOTE: this function assumes that the founder start position
of \code{founder_aln} is 1.}
}
\value{
Tibble with the same columns as output in \code{find_consensus()}, plus
an additional column, \code{conserved}, that indicates whether or not the position
is conserved (Yes means conserved, No means not conserved, NA means the
conserved position is a gap ('-'))
}
\description{
Identify conserved sites relative to the founder sequence.
}
\examples{
# this is a hack to get the data in the right format...
hiv_env_flt_2021 <- ape::as.matrix.DNAbin(hiv_env_flt_2021)
hxb2_founder <- ape::as.matrix.DNAbin(hxb2_founder)
identify_conserved_sites(hiv_env_flt_2021, 'B.FR.83.HXB2_LAI_IIIB_BRU.K03455')
identify_conserved_sites(hiv_env_flt_2021, 'B.US.2011.DEMB11US006.KC473833',
ref = 'B.FR.83.HXB2_LAI_IIIB_BRU.K03455', founder_aln = hxb2_founder)
}
