% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_wavess.R
\name{run_wavess}
\alias{run_wavess}
\title{Run wavess}
\usage{
run_wavess(
  pop_samp,
  founder_seqs,
  nt_sub_probs,
  prob_mut = 3.5e-05,
  prob_recomb = 1.4e-05,
  prob_act_to_lat = 0.001,
  prob_lat_to_act = 0.01,
  prob_lat_prolif = 0.01,
  prob_lat_die = 0.01,
  conserved_sites = NULL,
  conserved_cost = 0.99,
  ref_seq = NULL,
  rep_exp = 1,
  epitope_locations = NULL,
  seroconversion_time = 30,
  prop_for_imm = 0.01,
  gen_full_potency = 90,
  seed = NULL
)
}
\arguments{
\item{pop_samp}{Tibble with columns generation, active_cell_count, n_sample_active.
Can be generated using the \code{define_growth_curve()} and \code{define_sampling_scheme()} functions.}

\item{founder_seqs}{Founder sequences as a vector of character strings. For example c('ATCG', 'ATTT')}

\item{nt_sub_probs}{Named matrix of nucleotide substitution probabilities.
Rows are from, columns are to. Can be generated using the \code{calc_nt_subst_probs()} function.}

\item{prob_mut}{Probability of a mutation at one site in one generation
(default: 3.5e-5)}

\item{prob_recomb}{Probability of a recombination event at a given site in
one generation (default: 1.4e-5)}

\item{prob_act_to_lat}{Probability that an active cell becomes latent in a
generation (default: 0.001)}

\item{prob_lat_to_act}{Probability that a latent cell becomes active in a
generation (default: 0.01)}

\item{prob_lat_prolif}{Probability that a latent cell proliferates in a
generation (default: 0.01)}

\item{prob_lat_die}{Probability that a latent cell dies in a
generation (default: 0.01)}

\item{conserved_sites}{Vector of conserved sites.
This can be generated using the \code{identify_conserved_sites()} function
(default: NULL, i.e. no conserved sites fitness costs)}

\item{conserved_cost}{Cost of mutation at conserved site (default: 0.99)}

\item{ref_seq}{Reference sequence as a character string. A consensus sequence,
that can be used as the reference sequence, can be generated using the function
\code{find_consensus()} (default: NULL, i.e. no fitness cost relative to a reference sequence)}

\item{rep_exp}{Replicative fitness exponent, only relevant when ref_seq is not NULL (default: 1) # MAKE THIS CLEARER ONCE WE DECIDE ON A FINAL DEFINITION}

\item{epitope_locations}{Tibble of epitope locations and maximum fitness costs with columns
epi_start_nt, epi_end_nt, max_fitness_cost.
This can be generated using the functions \code{get_epitope_frequencies()} and \code{sample_epitopes()}
(default: NULL, i.e. no immune fitness costs)}

\item{seroconversion_time}{Generation at which seroconversion occurs, only
relevant when epitope_locations is not NULL (default: 30).}

\item{prop_for_imm}{Proportion of all infected cells that must be infected with
a given sequence for that sequence to be recognized by the immune system, only
relevant when epitope_locations is not NULL (default: 0.01).}

\item{gen_full_potency}{Number of generations it takes for an immune response
to an epitope to reach full potency, only relevant when epitope_locations is
not NULL (default: 90).}

\item{seed}{Optional seed (default: NULL)}
}
\value{
List including: tibble of counts, and alignment of sequences
}
\description{
Simulate within-host evolution.
Please note that the default arguments were set with the the HIV ENV gp120 gene
in mind. If you'd like to simulate something else, you may
have to modify certain parameters.
}
\examples{
\dontrun{
hiv_env_flt_2021 <- ape::as.matrix.DNAbin(hiv_env_flt_2021)
run_wavess(generate_pop_samp(gN = 300), c('ATCG', 'ATTT'),
calc_nt_sub_probs(hiv_env_flt_2021[1:3,]))
}
}
