% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_wavess.R
\name{run_wavess}
\alias{run_wavess}
\title{Run wavess}
\usage{
run_wavess(
  pop_samp,
  founder_seqs,
  nt_sub_probs,
  prob_mut = 3.5e-05,
  prob_recomb = 1.4e-05,
  prob_act_to_lat = 0.001,
  prob_lat_to_act = 0.01,
  prob_lat_prolif = 0.01,
  prob_lat_die = 0.01,
  conserved_sites = NULL,
  conserved_cost = 0.99,
  ref_seq = NULL,
  rep_exp = 1,
  epitope_locations = NULL,
  seroconversion_time = 30,
  prop_for_imm = 0.01,
  gen_full_potency = 90,
  seed = NULL
)
}
\arguments{
\item{pop_samp}{Tibble with columns generation, active_cell_count,
n_sample_active. Note that the initial active cell population size in the
first generation must be the same as the number of input founder sequences
(because it simply \emph{is} the input founder sequences). Can be generated
using the \code{\link[=generate_pop_samp]{generate_pop_samp()}} function.}

\item{founder_seqs}{Founder sequence(s) as a character string or a vector of
character strings, for example 'ACATG'. The founder sequence(s) may only
contain the characters ACGT, and no gaps are allowed. When modeling
immune fitness, they are expected to be codon-aligned.}

\item{nt_sub_probs}{Named matrix of nucleotide substitution probabilities.
Rows are from, columns are to. Can be generated using the
\code{\link[=calc_nt_sub_probs]{calc_nt_sub_probs()}} function.}

\item{prob_mut}{Probability of a mutation at one site in one generation
(default: 3.5e-5)}

\item{prob_recomb}{Probability of a recombination event at a given site in
one generation (default: 1.4e-5)}

\item{prob_act_to_lat}{Probability that an active cell becomes latent in a
generation (default: 0.001). Set this to 0 if you don't want to model
latent cell dynamics.}

\item{prob_lat_to_act}{Probability that a latent cell becomes active in a
generation (default: 0.01)}

\item{prob_lat_prolif}{Probability that a latent cell proliferates in a
generation (default: 0.01)}

\item{prob_lat_die}{Probability that a latent cell dies in a generation
(default: 0.01)}

\item{conserved_sites}{Vector of conserved sites. This can be generated using
the \code{\link[=identify_conserved_sites]{identify_conserved_sites()}} function (default: NULL, i.e. no conserved
sites fitness costs)}

\item{conserved_cost}{Cost of mutation at conserved site (default: 0.99)}

\item{ref_seq}{Reference sequence as a character string. A consensus
sequence, that can be used as the reference sequence, can be generated
using the function \code{\link[=identify_conserved_sites]{identify_conserved_sites()}} (default: NULL, i.e. no
fitness cost relative to a reference sequence)}

\item{rep_exp}{Replicative fitness exponent, only relevant when ref_seq is
not NULL (default: 1) # MAKE THIS CLEARER ONCE WE DECIDE ON A FINAL
DEFINITION}

\item{epitope_locations}{Tibble of epitope locations and maximum fitness
costs with columns epi_start_nt, epi_end_nt, max_fitness_cost. These
epitopes are expected to be in a protein in the correct reading frame, as
the nucleotide sequences are translated to amino acids to calculate the
immune fitness cost. This epitope location tibble can be generated using
the functions \code{\link[=get_epitope_frequencies]{get_epitope_frequencies()}} and \code{\link[=sample_epitopes]{sample_epitopes()}}.
(default: NULL, i.e. no immune fitness costs)}

\item{seroconversion_time}{Generation at which seroconversion occurs, only
relevant when epitope_locations is not NULL (default: 30).}

\item{prop_for_imm}{Proportion of all infected cells that must be infected
with a given sequence for that sequence to be recognized by the immune
system, only relevant when epitope_locations is not NULL (default: 0.01).}

\item{gen_full_potency}{Number of generations it takes for an immune response
to an epitope to reach full potency, only relevant when epitope_locations
is not NULL (default: 90).}

\item{seed}{Optional seed (default: NULL)}
}
\value{
List including: tibble of counts, and alignment of sampled sequences
}
\description{
Simulate within-host evolution optionally including recombination (default:
on), latency (default: on), and fitness costs (default: off). The three
fitness costs that can be simulated are conserved sites, fitness relative to
a reference sequence, and antibody-based immune fitness costs. Please note
that the default arguments were set with the the HIV ENV gp120 gene in mind.
If you'd like to simulate something else, you will likely have to modify
certain parameters. However, if you are interested in this gene in
particular, you can probably use most of the defaults including the founder
and reference sequences provided as examples. However, by default no fitness
costs are modeled. To model these, you can use the pre-processing functions
(see \code{vignette('prepare_input_data')}) to generate the relevant inputs. Also,
the parameters for latent probabilities are assumed to be small, such that it
is unlikely that multiple events (activate, die, proliferate) will occur to a
single latent cell in a single (active cell) generation. See
\code{vignette('run_wavess')} for more details about the simulator and input
arguments.
}
\examples{
\dontrun{
run_wavess(
  generate_pop_samp(n_gen = 300), "ATCG",
  calc_nt_sub_probs(hiv_env_flt_2022)
)
}
}
