% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_epitopes.R
\name{sample_epitopes}
\alias{sample_epitopes}
\title{Sample epitopes}
\usage{
sample_epitopes(
  epitope_probabilities,
  start_aa_pos = 0,
  end_aa_pos = NULL,
  num_epitopes = 10,
  aa_epitope_length = 10,
  max_fit_cost = 0.4,
  cost_type = "linear",
  max_resamples = 100,
  ref_founder_map = NULL
)
}
\arguments{
\item{epitope_probabilities}{Epitope probability tibble as output by
\code{\link[=get_epitope_frequencies]{get_epitope_frequencies()}}, including columns \code{aa_position} and
\code{epitope_probability}. \code{aa_position} should be indexed at 0}

\item{start_aa_pos}{Starting amino acid position to consider for epitopes,
indexed at 0 (default: 0, i.e. the first position)}

\item{end_aa_pos}{Ending amino acid position to consider for epitopes,
indexed at 0 (default: NULL, i.e. through the final position in
\code{epitope_probabilities$aa_position})}

\item{num_epitopes}{Number of epitopes to sample}

\item{aa_epitope_length}{Amino acid epitope length}

\item{max_fit_cost}{Maximum fitness cost of an epitope, must be between 0 and
1 (default: 0.4)
\strong{note that the model output is very sensitive to this parameter}}

\item{cost_type}{"linear" or "random"; linear returns max fitness costs
distributed linearly between 0 and \code{max_fit_cost} (not including 0, but
including \code{max_fit_cost}), random returns one epitope with \code{max_fit_cost}
and all other epitopes with max fitness costs randomly selected from a
uniform distribution between 0 and \code{max_fit_cost} (default: "linear")}

\item{max_resamples}{Maximum number of resampling events to attempt; this is
to prevent an infinite loop (default: 100)}

\item{ref_founder_map}{Output from \code{\link[=map_ref_founder]{map_ref_founder()}}, including
\emph{nucleotide} reference and founder positions (\code{ref_pos} and \code{founder_pos}).
\strong{NOTE:} The reference positions here, if they were converted to amino
acid positions, are expected to match with the reference positions in
\code{epitope_probabilities}. Further, we assume that the founder indices align
with the founder sequence positions to be used in the simulation (default:
NULL)}
}
\value{
tibble with the \code{num_epitopes} rows and the following columns:
\itemize{
\item \code{epi_start_nt}: nucleotide epitope start position
\item \code{epi_end_nt}: nucleotide epitope end position
\item \code{max_fitness_cost}: maximum fitness cost for that epitope
}
}
\description{
Sample epitopes based on epitope probabilities. Note that the positions
returned assume that the start of the amino acid sequence is also the start
of the founder sequence in the simulation. We also assume that there are no
frameshift mutations in the founder sequence.
}
\examples{
sample_epitopes(get_epitope_frequencies(env_features$Position - 1))
}
