% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_python_dependencies.R
\name{use_python_venv}
\alias{use_python_venv}
\title{Use wavess python virtual environment}
\usage{
use_python_venv(path = "inst/python")
}
\arguments{
\item{path}{Path to directory with agents.py}
}
\value{
functions in agents
}
\description{
Use wavess python virtual environment
}
