% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavess-package.R
\docType{package}
\name{wavess-package}
\alias{wavess}
\alias{wavess-package}
\title{wavess: Within-Host Agent-Based Viral Evolution Sequence Simulator}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Simulates within-host viral sequence evolution. Optionally includes recombination, a latent cell reservoir, and three types of selective pressures (conserved sites, comparison to a "most fit" sequence, and adaptive B-cell immunity).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://molevolepid.github.io/wavess/}
}

}
\author{
\strong{Maintainer}: Zena Lapp \email{zenalapp@lanl.gov} (\href{https://orcid.org/0000-0003-4674-2176}{ORCID})

Authors:
\itemize{
  \item Narmada Sambaturu \email{narmada.sambaturu@gmail.com} (\href{https://orcid.org/0000-0002-2570-3699}{ORCID})
}

Other contributors:
\itemize{
  \item Kenneth Benoit (\href{https://orcid.org/0000-0002-0797-564X}{ORCID}) (python install) [copyright holder]
  \item Akitaka Matsuo (\href{https://orcid.org/0000-0002-3323-6330}{ORCID}) (python install) [copyright holder]
  \item Johannes Gruber (\href{https://orcid.org/0000-0001-9177-1772}{ORCID}) (python install) [copyright holder]
}

}
\keyword{internal}
