% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavess-package.R
\docType{package}
\name{wavess-package}
\alias{wavess}
\alias{wavess-package}
\title{wavess: Within-Host Agent-Based Viral Evolution Sequence Simulator}
\description{
Simulates within-host viral sequence evolution. Optionally includes recombination, a latent cell reservoir, and three types of selective pressures (conserved sites, comparison to a "most fit" sequence, and adaptive B-cell immunity).
}
\author{
\strong{Maintainer}: Zena Lapp \email{zenalapp@lanl.gov} (\href{https://orcid.org/0000-0003-4674-2176}{ORCID})

Authors:
\itemize{
  \item Narmada Sambaturu \email{narmada.sambaturu@gmail.com} (\href{https://orcid.org/0000-0002-2570-3699}{ORCID})
}

Other contributors:
\itemize{
  \item Sophie Hoffman \email{shoffm@umich.edu} (\href{https://orcid.org/0000-0003-2518-6422}{ORCID}) (clusters) [copyright holder]
  \item Joyce Wang \email{wangjoy@umich.edu} (\href{https://orcid.org/0000-0002-8674-1015}{ORCID}) (clusters) [copyright holder]
  \item Evan Snitkin \email{esnitkin@umich.edu} (\href{https://orcid.org/0000-0001-8409-278X}{ORCID}) (clusters) [copyright holder]
}

}
\keyword{internal}
